requirements.txt
flask
psycopg2-binary
google-cloud-storage
google-cloud-secret-manager
functions-framework
